Move Tstamp implementation to the mathnum module.
