+++ src/lib/devices/Device.h
@@ -21,11 +21,11 @@
 #include <devices/Device_params.h>
 #include <kunquat/limits.h>
 #include <mathnum/Random.h>
+#include <mathnum/Tstamp.h>
 #include <player/Device_states.h>
 #include <player/Linear_controls.h>
 #include <player/Work_buffers.h>
 #include <string/Streader.h>
-#include <Tstamp.h>
 
 #include <stdbool.h>
 #include <stdint.h>
