+++ src/lib/devices/Device_field.h
@@ -23,8 +23,8 @@
 #include <devices/param_types/Sample.h>
 #include <devices/param_types/Sample_params.h>
 #include <mathnum/Real.h>
+#include <mathnum/Tstamp.h>
 #include <string/Streader.h>
-#include <Tstamp.h>
 
 #include <stdbool.h>
 #include <stdint.h>
