+++ src/lib/devices/Device_impl.h
@@ -23,10 +23,10 @@
 #include <devices/param_types/Note_map.h>
 #include <devices/param_types/Num_list.h>
 #include <devices/param_types/Sample.h>
+#include <mathnum/Tstamp.h>
 #include <player/Device_state.h>
 #include <player/Linear_controls.h>
 #include <string/key_pattern.h>
-#include <Tstamp.h>
 
 #include <stdbool.h>
 #include <stdint.h>
