+++ src/lib/mathnum/Tstamp.c
@@ -12,7 +12,7 @@
  */
 
 
-#include <Tstamp.h>
+#include <mathnum/Tstamp.h>
 
 #include <debug/assert.h>
