+++ src/lib/module/sheet/Column.h
@@ -17,10 +17,10 @@
 
 
 #include <containers/AAtree.h>
+#include <mathnum/Tstamp.h>
 #include <module/sheet/Trigger.h>
 #include <player/Event_names.h>
 #include <string/Streader.h>
-#include <Tstamp.h>
 
 #include <stdbool.h>
 #include <stdint.h>
