+++ src/lib/module/sheet/Trigger.h
@@ -17,10 +17,10 @@
 
 
 #include <kunquat/limits.h>
+#include <mathnum/Tstamp.h>
 #include <player/Event_names.h>
 #include <player/Event_type.h>
 #include <string/Streader.h>
-#include <Tstamp.h>
 
 #include <stdlib.h>
