+++ src/lib/player/Active_jumps.h
@@ -18,10 +18,10 @@
 
 #include <containers/AAtree.h>
 #include <kunquat/limits.h>
+#include <mathnum/Tstamp.h>
 #include <player/Jump_cache.h>
 #include <player/Jump_context.h>
 #include <Pat_inst_ref.h>
-#include <Tstamp.h>
 
 #include <stdlib.h>
