+++ src/lib/player/Position.h
@@ -16,8 +16,8 @@
 #define K_POSITION_H
 
 
+#include <mathnum/Tstamp.h>
 #include <Pat_inst_ref.h>
-#include <Tstamp.h>
 
 #include <stdbool.h>
 #include <stdint.h>
