+++ src/lib/string/Streader.h
@@ -17,8 +17,8 @@
 
 
 #include <Error.h>
+#include <mathnum/Tstamp.h>
 #include <Pat_inst_ref.h>
-#include <Tstamp.h>
 
 #include <stdbool.h>
 #include <stdint.h>
