+++ src/lib/string/serialise.h
@@ -13,8 +13,8 @@
 
 
 #include <mathnum/Real.h>
+#include <mathnum/Tstamp.h>
 #include <Pat_inst_ref.h>
-#include <Tstamp.h>
 
 #include <stdbool.h>
 #include <stdint.h>
